# apps/knot-space/swagger.yaml
openapi: 3.0.0
info:
  title: Knot Space API
  version: 0.1.0
  description: Online repository server for Knot packages
  contact:
    name: Knot Team
    url: https://github.com/saravenpi/knot
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.knot.space
    description: Production server
...
paths:
...
  /api/users/{username}:
    get:
      tags:
        - auth
      summary: Get user by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          description: User not found
  /api/teams:
    post:
      tags:
        - teams
      summary: Create a new team
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '200':
          description: Team created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          description: Invalid input
        '409':
          description: Team already exists

    get:
      tags:
        - teams
      summary: List all teams
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
  /api/teams/{teamName}:
    get:
      tags:
        - teams
      summary: Get team by name
      parameters:
        - name: teamName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Team details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '404':
          description: Team not found

    delete:
      tags:
        - teams
      summary: Delete team (owner only)
      security:
        - BearerAuth: []
      parameters:
        - name: teamName
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Team deleted successfully
        '403':
          description: Forbidden
        '404':
          description: Team not found
  /api/teams/{teamName}/members:
    get:
      tags:
        - teams
      summary: Get team members
      parameters:
        - name: teamName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    role:
                      type: string
                    joined_at:
                      type: string
                      format: date-time
                    user:
                      $ref: '#/components/schemas/UserProfile'

    post:
      tags:
        - teams
      summary: Add member to team
      security:
        - BearerAuth: []
      parameters:
        - name: teamName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTeamMemberRequest'
      responses:
        '201':
          description: Member added successfully
        '403':
          description: Forbidden
        '404':
          description: Team or user not found
        '409':
          description: User already in team
  /api/teams/{teamName}/members/{username}:
    delete:
      tags:
        - teams
      summary: Remove member from team
      security:
        - BearerAuth: []
      parameters:
        - name: teamName
          in: path
          required: true
          schema:
            type: string
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Member removed successfully
        '403':
          description: Forbidden
        '404':
          description: Team or user not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
...
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTeamRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
        description:
          type: string
          nullable: true
...
    AddTeamMemberRequest:
      type: object
      required:
        - username
        - role
      properties:
        username:
          type: string
        role:
          type: string
          enum: [admin, member]
...